Fix potentially tragic error with geocoding validation
geopy and GoogleV3 geocoders appear to explode if passed None or an
empty string.  Both of which were previously guarded, but were removed
during this evolution.  As a result, the wrapped geocoder caused the
entirety of bulk imports to freeze.  As a side note, return statements
are required in validation, otherwise code will also freeze.